Refactored UX to be modularized better
