--- resources/gaming-fe/src/hooks/useWasmBlob.ts
@@ -1,584 +1,0 @@
-import { useState, useEffect, useCallback, useRef } from 'react';
-import { Program } from 'clvm-lib';
-import { toast } from 'sonner';
-import {
-  storeInitArgs,
-} from './WasmStateInit';
-import {
-  GameConnectionState,
-  CalpokerOutcome,
-  BlockchainInboundAddressResult,
-  BlockchainReport,
-  OutcomeLogLine,
-  handValueToDescription,
-  WasmEvent,
-} from '../types/ChiaGaming';
-import { ChildFrameBlockchainInterface } from './ChildFrameBlockchainInterface';
-import {
-  getBlobSingleton,
-  initStarted,
-  setInitStarted,
-} from './blobSingleton';
-import { WasmBlobWrapper } from './WasmBlobWrapper';
-import { toHexString } from '../util';
-
-function coinPayloadToHex(coin: unknown): string | undefined {
-  if (Array.isArray(coin) && coin.length > 0 && coin.every((b): b is number => typeof b === 'number')) {
-    return toHexString(coin);
-  }
-  return undefined;
-}
-
-const TERMINAL_TYPES = [
-  'WeTimedOut', 'OpponentTimedOut', 'WeSlashedOpponent',
-  'OpponentSlashedUs', 'OpponentSuccessfullyCheated',
-  'GameCancelled', 'GameError', 'ChannelError',
-];
-
-function isTerminal(n: any): boolean {
-  return TERMINAL_TYPES.some(t => t in n);
-}
-
-function parseCards(readableBytes: number[], iStarted: boolean): { playerHand: number[], opponentHand: number[] } {
-  const program = Program.deserialize(Uint8Array.from(readableBytes));
-  const card_lists = program.toList().map(l => l.toList().map(v => v.toInt()));
-  if (iStarted) {
-    return { playerHand: card_lists[1], opponentHand: card_lists[0] };
-  } else {
-    return { playerHand: card_lists[0], opponentHand: card_lists[1] };
-  }
-}
-
-function selectedCardsToBitfield(selectedCards: number[], hand: number[]): number {
-  let bitfield = 0;
-  hand.forEach((cardId, index) => {
-    if (selectedCards.includes(cardId)) {
-      bitfield |= 1 << index;
-    }
-  });
-  return bitfield;
-}
-
-export interface UseWasmBlobResult {
-  error: any;
-  log: OutcomeLogLine[];
-  amount: number;
-  addressData: BlockchainInboundAddressResult | undefined;
-  ourShare: number | undefined;
-  theirShare: number | undefined;
-  gameConnectionState: GameConnectionState;
-  isPlayerTurn: boolean;
-  iStarted: boolean;
-  moveNumber: number;
-  handleMakeMove: () => void;
-  handleCheat: () => void;
-  playerHand: number[];
-  opponentHand: number[];
-  playerNumber: number;
-  cardSelections: number[];
-  setCardSelections: (s: number[] | ((prev: number[]) => number[])) => void;
-  outcome: CalpokerOutcome | undefined;
-  lastOutcome: CalpokerOutcome | undefined;
-  playAgain: () => void;
-  stopPlaying: () => void;
-  channelCoinHex: string | undefined;
-  channelStatus: string | undefined;
-  gameCoinHex: string | undefined;
-  gameStatus: string | undefined;
-}
-
-export function useWasmBlob(searchParams: any, lobbyUrl: string, uniqueId: string): UseWasmBlobResult {
-  const [ourShare, setOurShare] = useState<number | undefined>(undefined);
-  const [theirShare, setTheirShare] = useState<number | undefined>(undefined);
-  const [gameConnectionState, setGameConnectionState] =
-    useState<GameConnectionState>({
-      stateIdentifier: 'starting',
-      stateDetail: ['before handshake'],
-    });
-
-  const iStarted = searchParams.iStarted !== 'false';
-  const playerNumber = iStarted ? 1 : 2;
-  const [log, setLog] = useState<OutcomeLogLine[]>([]);
-  const [addressData, setAddressData] =
-    useState<BlockchainInboundAddressResult | undefined>(undefined);
-  const [playerHand, setPlayerHand] = useState<number[]>([]);
-  const [opponentHand, setOpponentHand] = useState<number[]>([]);
-  const [outcome, setOutcome] = useState<CalpokerOutcome | undefined>(undefined);
-  const [lastOutcome, setLastOutcome] = useState<CalpokerOutcome | undefined>(undefined);
-  const [isPlayerTurn, setMyTurn] = useState<boolean>(false);
-  const [moveNumber, setMoveNumber] = useState<number>(0);
-  const [error, setRealError] = useState<string | undefined>(undefined);
-  const [cardSelections, setOurCardSelections] = useState<number[]>([]);
-  const [gameIds, setGameIds] = useState<string[]>([]);
-  const [channelCoinHex, setChannelCoinHex] = useState<string | undefined>(undefined);
-  const [channelStatus, setChannelStatus] = useState<string | undefined>(undefined);
-  const [gameCoinHex, setGameCoinHex] = useState<string | undefined>(undefined);
-  const [gameStatus, setGameStatus] = useState<string | undefined>(undefined);
-  const amount = parseInt(searchParams.amount);
-
-  const playerHandRef = useRef<number[]>([]);
-  const opponentHandRef = useRef<number[]>([]);
-  const cardSelectionsRef = useRef<number[]>([]);
-  const moveNumberRef = useRef<number>(0);
-  const gameIdsRef = useRef<string[]>([]);
-  const gameOutcomeRef = useRef<CalpokerOutcome | undefined>(undefined);
-  const pendingProposalIdRef = useRef<string | null>(null);
-  const wantsNewGameRef = useRef<boolean>(false);
-  const firstGameAcceptedRef = useRef<boolean>(false);
-
-  playerHandRef.current = playerHand;
-  opponentHandRef.current = opponentHand;
-  cardSelectionsRef.current = cardSelections;
-  moveNumberRef.current = moveNumber;
-  gameIdsRef.current = gameIds;
-
-  const setError = (e: any) => {
-    if (e !== undefined) {
-      setRealError((prev) => prev === undefined ? e : prev);
-    }
-  };
-
-  let perGameAmount = amount / 10;
-  try {
-    perGameAmount = parseInt(searchParams.perGame);
-  } catch (e) {
-    if (searchParams.perGame) {
-      throw e;
-    }
-  }
-
-  const blockchain = new ChildFrameBlockchainInterface();
-
-  const { gameObject } = getBlobSingleton(
-    blockchain,
-    searchParams,
-    lobbyUrl,
-    uniqueId,
-    amount,
-    iStarted,
-  );
-
-  const gameObjectRef = useRef<WasmBlobWrapper>(gameObject);
-  gameObjectRef.current = gameObject;
-
-  const proposeNewGame = useCallback(() => {
-    const go = gameObjectRef.current;
-    if (!go || !go.isChannelReady()) return;
-    try {
-      go.proposeGame({
-        game_type: '63616c706f6b6572',
-        timeout: 100,
-        amount: perGameAmount,
-        my_contribution: perGameAmount / 2,
-        my_turn: !iStarted,
-        parameters: null,
-      });
-      console.log('[calpoker] proposed game');
-    } catch (e) {
-      console.error('proposeGame failed:', e);
-    }
-  }, [iStarted, perGameAmount]);
-
-  const recognizeOutcome = useCallback((newOutcome: CalpokerOutcome | undefined) => {
-    setOutcome(newOutcome);
-    gameOutcomeRef.current = newOutcome;
-    if (newOutcome) {
-      const iAmAlice = !iStarted;
-      const mySelects = iAmAlice ? newOutcome.alice_selects : newOutcome.bob_selects;
-      const theirSelects = iAmAlice ? newOutcome.bob_selects : newOutcome.alice_selects;
-      const myFinalHand = iAmAlice ? newOutcome.alice_final_hand : newOutcome.bob_final_hand;
-      const opponentFinalHand = iAmAlice ? newOutcome.bob_final_hand : newOutcome.alice_final_hand;
-      const myCards = iAmAlice ? newOutcome.alice_used_cards : newOutcome.bob_used_cards;
-      const myValue = iAmAlice ? newOutcome.alice_hand_value : newOutcome.bob_hand_value;
-      const theirCards = iAmAlice ? newOutcome.bob_used_cards : newOutcome.alice_used_cards;
-      const theirValue = iAmAlice ? newOutcome.bob_hand_value : newOutcome.alice_hand_value;
-      const myHandDescription = handValueToDescription(myValue, myCards);
-      const opponentHandDescription = handValueToDescription(theirValue, theirCards);
-      const newLogObject: OutcomeLogLine = {
-        topLineOutcome: newOutcome.my_win_outcome,
-        myHandDescription,
-        opponentHandDescription,
-        myHand: myCards,
-        opponentHand: theirCards,
-        myStartHand: playerHandRef.current,
-        opponentStartHand: opponentHandRef.current,
-        myFinalHand,
-        opponentFinalHand,
-        mySelects,
-        opponentSelects: theirSelects,
-        myPicks: iAmAlice ? newOutcome.alice_discards : newOutcome.bob_discards,
-        opponentPicks: iAmAlice ? newOutcome.bob_discards : newOutcome.alice_discards,
-      };
-      setLog(prev => [newLogObject, ...prev]);
-    }
-  }, [iStarted]);
-
-  const handleNotification = useCallback((n: any) => {
-    const go = gameObjectRef.current;
-    if (typeof n !== 'object' || n === null) return;
-
-    // Show toasts for notable game events. Channel vs game IDs so each category replaces its previous toast.
-    const type = Object.keys(n)[0];
-    const p = n[type]; // payload
-    const CHANNEL_TOAST_ID = 'channel-notification';
-    const GAME_TOAST_ID = 'game-notification';
-    type ToastCfg = { title: string; description?: string; variant?: 'default' | 'destructive'; toastId: string };
-    const channelToasts: Record<string, Omit<ToastCfg, 'toastId'>> = {
-      GoingOnChain:                { variant: 'default',     title: 'Going On-Chain',              description: p?.reason ?? 'Dispute detected — submitting to blockchain' },
-      ChannelCoinSpent:            { variant: 'default',     title: 'Channel Coin Spent',           description: 'The state channel coin was spent on-chain' },
-      UnrollCoinSpent:             { variant: 'default',     title: 'Unroll Coin Spent',            description: p?.reward_coin ? 'Unroll resolved — reward coin received' : 'The unroll coin was spent on-chain' },
-      StaleChannelUnroll:          { variant: 'destructive', title: 'Stale Channel Unrolled',       description: p?.our_reward !== undefined ? `You received ${p.our_reward} mojos` : 'Opponent\'s stale unroll resolved on-chain' },
-      ChannelError:                { variant: 'destructive', title: 'Channel Error',                description: p?.reason },
-      CleanShutdownStarted:        { variant: 'default',     title: 'Session Ending',               description: 'Opponent initiated a clean shutdown' },
-      CleanShutdownComplete:       { variant: 'default',     title: 'Session Ended',                description: 'Channel closed — funds returned on-chain' },
-    };
-    const gameToasts: Record<string, Omit<ToastCfg, 'toastId'>> = {
-      OpponentPlayedIllegalMove:   { variant: 'default',     title: 'Illegal Move Detected',       description: `Game #${p?.id} — slashing opponent on-chain…` },
-      WeSlashedOpponent:           { variant: 'default',     title: 'Opponent Slashed!',            description: `Game #${p?.id} — successfully claimed all game funds` },
-      OpponentSlashedUs:           { variant: 'destructive', title: 'You Were Slashed',             description: `Game #${p?.id} — your illegal move was proven on-chain` },
-      OpponentSuccessfullyCheated: { variant: 'destructive', title: 'Opponent Got Away',            description: p?.our_reward !== undefined ? `Game #${p?.id} — slash window expired, you received ${p.our_reward} mojos` : `Game #${p?.id} — slash window expired` },
-      WeTimedOut:                  { variant: 'destructive', title: 'You Timed Out',                description: p?.our_reward !== undefined ? `Game #${p?.id} — you received ${p.our_reward} mojos` : `Game #${p?.id}` },
-      OpponentTimedOut:            { variant: 'default',     title: 'Opponent Timed Out',           description: p?.our_reward !== undefined ? `Game #${p?.id} — you received ${p.our_reward} mojos` : `Game #${p?.id}` },
-      GameCancelled:               { variant: 'default',     title: 'Game Cancelled',               description: `Game #${p?.id} was cancelled` },
-      GameProposalCancelled:       { variant: 'destructive', title: 'Game Proposal Cancelled',      description: p?.reason ? `Game #${p?.id} — ${p.reason}` : `Game #${p?.id}` },
-      InsufficientBalance:         { variant: 'destructive', title: 'Insufficient Balance',         description: p?.our_balance_short && p?.their_balance_short ? 'Both sides have insufficient balance' : p?.our_balance_short ? 'Your balance is too low for this game' : 'Opponent\'s balance is too low for this game' },
-      GameError:                   { variant: 'destructive', title: 'Game Error',                   description: p?.reason ? `Game #${p?.id} — ${p.reason}` : `Game #${p?.id}` },
-      GameOnChain:                 { variant: 'default',     title: 'Game On-Chain',                description: p?.coin ? `Game #${p?.id} — coin on-chain` : `Game #${p?.id}` },
-    };
-    const tChannel = channelToasts[type];
-    const tGame = gameToasts[type];
-    const t = tChannel ?? tGame;
-    if (t) {
-      // Game-level notifications that carry an id must match the active game; otherwise escalate to channel error.
-      const activeGameId = gameIdsRef.current[0];
-      const notificationGameId = p?.id != null ? String(p.id) : null;
-      if (tGame && notificationGameId != null && activeGameId != null && notificationGameId !== activeGameId) {
-        toast.error(`Unexpected game notification for #${notificationGameId}`, { id: CHANNEL_TOAST_ID });
-      } else {
-        const toastId = tChannel ? CHANNEL_TOAST_ID : GAME_TOAST_ID;
-        const opts = { ...(t.description ? { description: t.description } : {}), id: toastId };
-        if (t.variant === 'destructive') {
-          toast.error(t.title, opts);
-        } else {
-          toast(t.title, opts);
-        }
-      }
-    }
-
-    // Update channel/game coin state for header and bar display (Task 4)
-    if ('ChannelCreated' in n) {
-      const hex = coinPayloadToHex(n.ChannelCreated?.channel_coin);
-      if (hex != null) setChannelCoinHex(hex);
-      setChannelStatus('Open');
-    } else if ('ChannelCoinSpent' in n) {
-      const hex = coinPayloadToHex(n.ChannelCoinSpent?.unroll_coin);
-      if (hex != null) setChannelCoinHex(hex);
-      setChannelStatus('Channel spent');
-    } else if ('UnrollCoinSpent' in n) {
-      const hex = coinPayloadToHex(n.UnrollCoinSpent?.reward_coin);
-      if (hex != null) setChannelCoinHex(hex);
-      setChannelStatus('Unroll spent');
-    }
-    if ('GameOnChain' in n) {
-      const hex = coinPayloadToHex(n.GameOnChain?.coin);
-      if (hex != null) setGameCoinHex(hex);
-      const ourTurn = n.GameOnChain?.our_turn;
-      setGameStatus(ourTurn === true ? 'Our turn' : ourTurn === false ? 'Their turn' : 'On-chain');
-    }
-
-    if ('GameProposed' in n) {
-      if (!iStarted) {
-        const proposalId = n.GameProposed.id.toString();
-        if (!firstGameAcceptedRef.current || wantsNewGameRef.current) {
-          // First game: auto-accept. Or player already clicked Play Again.
-          wantsNewGameRef.current = false;
-          try {
-            go?.acceptProposal(proposalId);
-          } catch (e) {
-            console.error('acceptProposal failed:', e);
-          }
-        } else {
-          // Subsequent game: stash until player clicks Play Again.
-          pendingProposalIdRef.current = proposalId;
-        }
-      }
-    } else if ('GameProposalAccepted' in n) {
-      firstGameAcceptedRef.current = true;
-      const newId = n.GameProposalAccepted.id.toString();
-      setGameIds(prev => [...prev, newId]);
-      gameIdsRef.current = [...gameIdsRef.current, newId];
-      // Alice (joiner, iStarted=false) moves first at move 0
-      setMyTurn(!iStarted);
-      setMoveNumber(0);
-      moveNumberRef.current = 0;
-      setPlayerHand([]);
-      setOpponentHand([]);
-      playerHandRef.current = [];
-      opponentHandRef.current = [];
-      setOurCardSelections([]);
-      cardSelectionsRef.current = [];
-      setOutcome(undefined);
-      setGameConnectionState({ stateIdentifier: 'running', stateDetail: [] });
-    } else if ('OpponentMoved' in n) {
-      const currentMove = moveNumberRef.current;
-
-      setMyTurn(true);
-
-      if (currentMove === 1 && !iStarted) {
-        try {
-          const cards = parseCards(n.OpponentMoved.readable, iStarted);
-          setPlayerHand(cards.playerHand);
-          setOpponentHand(cards.opponentHand);
-          playerHandRef.current = cards.playerHand;
-          opponentHandRef.current = cards.opponentHand;
-        } catch (e) {
-          console.error('parseCards from OpponentMoved failed:', e);
-        }
-      } else if (currentMove >= 2) {
-        const myDiscardsBitfield = selectedCardsToBitfield(
-          cardSelectionsRef.current,
-          playerHandRef.current,
-        );
-        const newOutcome = new CalpokerOutcome(
-          iStarted,
-          myDiscardsBitfield,
-          iStarted ? opponentHandRef.current : playerHandRef.current,
-          iStarted ? playerHandRef.current : opponentHandRef.current,
-          n.OpponentMoved.readable,
-        );
-        recognizeOutcome(newOutcome);
-        const gameId = gameIdsRef.current[0];
-        // Reset wasm-facing refs immediately so no stale calls can happen.
-        setGameIds(prev => prev.slice(1));
-        gameIdsRef.current = gameIdsRef.current.slice(1);
-        setMyTurn(false);
-        setOurCardSelections([]);
-        cardSelectionsRef.current = [];
-        setLastOutcome(gameOutcomeRef.current);
-
-        if (!iStarted && currentMove === 2) {
-          // Alice: send final reveal move, then reset immediately.
-          // CaliforniaPoker guards against resetting while animation is running.
-          try {
-            go?.makeMove(gameId, null);
-          } catch (e) {
-            console.error('makeMove failed:', e);
-          }
-          setMoveNumber(0);
-          moveNumberRef.current = 0;
-          setPlayerHand([]);
-          setOpponentHand([]);
-          playerHandRef.current = [];
-          opponentHandRef.current = [];
-        } else {
-          // Bob: reset immediately. New game is proposed when he clicks Play Again.
-          setMoveNumber(0);
-          moveNumberRef.current = 0;
-          setPlayerHand([]);
-          setOpponentHand([]);
-          playerHandRef.current = [];
-          opponentHandRef.current = [];
-        }
-      }
-    } else if ('GameMessage' in n) {
-      try {
-        const cards = parseCards(n.GameMessage.readable, iStarted);
-        setPlayerHand(cards.playerHand);
-        setOpponentHand(cards.opponentHand);
-        playerHandRef.current = cards.playerHand;
-        opponentHandRef.current = cards.opponentHand;
-      } catch (e) {
-        console.error('parseCards failed:', e, 'readable:', n.GameMessage.readable);
-      }
-    } else if ('CleanShutdownComplete' in n) {
-      setGameConnectionState({ stateIdentifier: 'clean_shutdown', stateDetail: [] });
-    } else if ('ChannelCreated' in n) {
-      setGameConnectionState({ stateIdentifier: 'running', stateDetail: [] });
-      if (iStarted) {
-        proposeNewGame();
-      }
-    } else if ('CleanShutdownStarted' in n) {
-      // Peer initiated clean shutdown
-    } else if ('GoingOnChain' in n) {
-      setGameConnectionState({ stateIdentifier: 'running', stateDetail: ['On-chain dispute in progress'] });
-    } else if (isTerminal(n)) {
-      // Only reset game-specific UI state when the game is still tracked.
-      // With auto-accept, WeTimedOut/OpponentTimedOut fire for games that
-      // were already cleaned up by the OpponentMoved handler; redundant
-      // resets cause CaliforniaPoker's cards to jump (pre-swap fallback).
-      const hadActiveGame = gameIdsRef.current.length > 0;
-      if (hadActiveGame) {
-        setGameIds(prev => prev.slice(1));
-        gameIdsRef.current = gameIdsRef.current.slice(1);
-        setMyTurn(false);
-        setOurCardSelections([]);
-        cardSelectionsRef.current = [];
-        setMoveNumber(0);
-        moveNumberRef.current = 0;
-        setPlayerHand([]);
-        setOpponentHand([]);
-        playerHandRef.current = [];
-        opponentHandRef.current = [];
-        setOutcome(undefined);
-        setGameConnectionState({ stateIdentifier: 'running', stateDetail: [] });
-      }
-      setLastOutcome(gameOutcomeRef.current);
-
-      // New game is proposed when the player clicks Play Again.
-    } else {
-      console.warn('unhandled notification:', JSON.stringify(n));
-    }
-  }, [iStarted, recognizeOutcome, proposeNewGame]);
-
-  useEffect(() => {
-    const subscription = gameObject.getObservable().subscribe({
-      next: (evt: WasmEvent) => {
-        switch (evt.type) {
-          case 'notification':
-            handleNotification(evt.data);
-            break;
-          case 'error':
-            setError(evt.error);
-            break;
-          case 'finished':
-            setGameConnectionState({ stateIdentifier: 'clean_shutdown', stateDetail: [] });
-            break;
-          case 'address':
-            setAddressData(evt.data);
-            break;
-          default:
-            console.warn('unhandled event type:', (evt as any).type, evt);
-            break;
-        }
-      }
-    });
-
-    if (!initStarted) {
-      setInitStarted(true);
-    }
-
-    return () => {
-      subscription.unsubscribe();
-    };
-  }, [gameObject, iStarted, handleNotification, proposeNewGame]);
-
-  useEffect(() => {
-    const subscription = blockchain.getObservable().subscribe({
-      next: (e: BlockchainReport) => {
-        gameObject?.blockNotification(e.peak, e.block, e.report);
-      },
-    });
-
-    return () => {
-      subscription.unsubscribe();
-    };
-  }, [gameObject]);
-
-  const handleMakeMove = useCallback(() => {
-    const go = gameObjectRef.current;
-    if (!go || !go.isChannelReady()) return;
-    const currentGameId = gameIdsRef.current[0];
-    if (!currentGameId) return;
-
-    const currentMove = moveNumberRef.current;
-
-    if (currentMove === 0) {
-      go.makeMove(currentGameId, null);
-      const newMoveNum = currentMove + 1;
-      setMoveNumber(newMoveNum);
-      moveNumberRef.current = newMoveNum;
-      setMyTurn(false);
-    } else if (currentMove === 1) {
-      if (cardSelectionsRef.current.length !== 4) return;
-      const cards = cardSelectionsRef.current;
-      go.makeMove(currentGameId, Program.fromList(cards.map(c => Program.fromInt(c))));
-      const newMoveNum = currentMove + 1;
-      setMoveNumber(newMoveNum);
-      moveNumberRef.current = newMoveNum;
-      setMyTurn(false);
-    } else if (currentMove === 2) {
-      go.makeMove(currentGameId, null);
-      const newMoveNum = currentMove + 1;
-      setMoveNumber(newMoveNum);
-      moveNumberRef.current = newMoveNum;
-      setMyTurn(false);
-    }
-  }, []);
-
-  const handleCheat = useCallback(() => {
-    const go = gameObjectRef.current;
-    const currentGameId = gameIdsRef.current[0];
-    if (!go || !currentGameId) return;
-    go.cheat(currentGameId, 0);
-  }, []);
-
-  const setCardSelections = useCallback((selectionsOrFn: number[] | ((prev: number[]) => number[])) => {
-    if (typeof selectionsOrFn === 'function') {
-      setOurCardSelections(prev => {
-        const next = selectionsOrFn(prev);
-        cardSelectionsRef.current = next;
-        return next;
-      });
-    } else {
-      setOurCardSelections(selectionsOrFn);
-      cardSelectionsRef.current = selectionsOrFn;
-    }
-  }, []);
-
-  const playAgain = useCallback(() => {
-    if (iStarted) {
-      // Bob: propose a new game now that the user has opted in.
-      proposeNewGame();
-    } else {
-      // Alice: accept the pending proposal, or flag that we want the next one.
-      const pending = pendingProposalIdRef.current;
-      if (pending) {
-        pendingProposalIdRef.current = null;
-        try {
-          gameObjectRef.current?.acceptProposal(pending);
-        } catch (e) {
-          console.error('acceptProposal failed:', e);
-        }
-      } else {
-        wantsNewGameRef.current = true;
-      }
-    }
-  }, [iStarted, proposeNewGame]);
-
-  const stopPlaying = useCallback(() => {
-    gameObject?.cleanShutdown();
-  }, [gameObject]);
-
-  (window as any).loadWasm = useCallback((chia_gaming_init: any, cg: any) => {
-    storeInitArgs(chia_gaming_init, cg);
-  }, []);
-
-  return {
-    error,
-    addressData: addressData || { address: '', puzzleHash: '' },
-    amount,
-    ourShare,
-    theirShare,
-    log,
-    gameConnectionState,
-    isPlayerTurn,
-    iStarted,
-    playerNumber,
-    handleMakeMove,
-    handleCheat,
-    playerHand,
-    opponentHand,
-    moveNumber,
-    cardSelections,
-    setCardSelections,
-    playAgain,
-    stopPlaying,
-    outcome,
-    lastOutcome,
-    channelCoinHex,
-    channelStatus,
-    gameCoinHex,
-    gameStatus,
-  };
-}
-
